refactor: align Talent Pioneers page with site standards
